Successfully filter trips and display trip cards to dom
