+++ src/Traveler.js
@@ -18,13 +18,15 @@ class Traveler {
     let filteredTrips = trips.filter(trip => trip.userID === this.id);
     filteredTrips.forEach(trip => {
       destinations.forEach(location => {
+        if (trip.destinationID === location.id) {
           currentDestination = location;
           this.allTrips.push(new Trip(trip, currentDestination));
+        }
       })
     })
     return this.allTrips
   }
-  
+
   getUserName() {
   if (!this.name) {
     return 'Hi, Friend! Please Add Your Name To Your User Profile!'
