+++ src/domUpdates.js
@@ -15,15 +15,20 @@ const displayYearlySpending = (total) => {
 }
 
 const invalidLogin = () => {
-  console.log()
   loginMessage.innerText = 'Invalid Login and/or Password. Try Again.'
 }
 
-const displayTrips = (currentUser) => {
+const loginSubmit = () => {
+  show(mainPage)
+  hide(loginPage)
+}
+
+const displayTrips = (traveler) => {
+  console.log(traveler)
     tripCards.innerHTML = '';
     let tripInfo = '';
-    if (currentUser.allTrips.length > 0) {
-      currentUser.allTrips.forEach(trip => {
+    if (traveler.allTrips.length > 0) {
+      traveler.allTrips.forEach(trip => {
         // const formattedDate = this.formatDate(trip.date);
         tripInfo += `
         <article class="trip-cards">
@@ -31,7 +36,7 @@ const displayTrips = (currentUser) => {
         <img class="trip-img" src=${trip.destination.image} alt=${trip.destination.alt}>
         </div>
         <h3 class="destination-name">${trip.destination.destination}</h3>
-        <p>Trip date: ${formattedDate} <br>
+        <p>Trip date:  <br>
         Travelers: ${trip.travelerCount} <br>
         Duration: ${trip.duration} <br>
         Status: ${trip.status} <br> </p>
@@ -47,4 +52,11 @@ const displayTrips = (currentUser) => {
     tripCards.insertAdjacentHTML('beforeend', tripInfo);
   }
 
-export {greetUser, displayYearlySpending, displayTrips, invalidLogin}
+  const hide = (section) => {
+    section.classList.toggle('hidden')
+  }
+
+  const show = (section) => {
+    section.classList.toggle('hidden')
+  }
+export {greetUser, displayYearlySpending, displayTrips, invalidLogin, loginSubmit}
